requirements.txt
pandas
numpy
yfinance
pytz
requests
mysql-connector-python
cryptography
python-dotenv